requirements.txt
streamlit
pillow
pytesseract
langchain
openai
python-dotenv
langchain-google-genai
